Improve JavascriptExecution message registration
